Fix 'DOCTYPE is disallowed' in build script
Using gradle 2.1 the build script fails with the following message:

    [Fatal Error] SerialDateWidget.gwt.xml:2:10: DOCTYPE is disallowed when the feature "http://apache.org/xml/features/disallow-doctype-decl" set to true.

The build script uses the default XmlParser. This is 'a non-validating and non-namespace-aware XmlParser which does not allow DOCTYPE declarations in documents'. Correcting it to use a parser that does allow DOCTYPE declarations still tries to fetch an external DTD and fails with a:

    [Fatal Error] SerialDateWidget.gwt.xml:2:172: External DTD: Failed to read external DTD 'gwt-module.dtd', because 'http' access is not allowed due to restriction set by the accessExternalDTD property.


Setting the `load-external-dtd` feature to false fixes the failure.